move helper function to separate file
