+++ app.py
@@ -3,6 +3,7 @@ import streamlit as st
 import pandas as pd 
 from mitosheet.streamlit.v1 import spreadsheet, RunnableAnalysis
 import os
+from custom_functions import CHECK
 from file_utils import save_analysis, get_all_provider_names_from_automations, get_runnable_analysis_and_description_from_provider_name
 
 # Set the streamlit page to wide so you can see the whole spreadsheet
@@ -13,29 +14,6 @@ if not os.path.exists(os.path.join(os.getcwd(), 'automations')):
 
 if not os.path.exists(os.path.join(os.getcwd(), 'data')):
     os.mkdir(os.path.join(os.getcwd(), 'data'))
-
-def CHECK(val1, val2, mismatch_label=False):
-
-    # For each cell in val 1 and val2 perform the following check
-
-    def _check_if_value_is_none(val):
-        if val is None or pd.isna(val):
-            return True
-        else:
-            return False
-    
-    def _check(val1, val2, mismatch_label):
-
-        if _check_if_value_is_none(val1) and _check_if_value_is_none(val2):
-            return True if not mismatch_label else ''
-        elif val1 == val2:
-            return True if not mismatch_label else ''
-        else:
-            return mismatch_label
-
-    # Use the _check on series val1 and val2
-    if isinstance(val1, pd.Series) and isinstance(val2, pd.Series):
-        return pd.Series([_check(v1, v2, mismatch_label) for v1, v2 in zip(val1, val2)])
 
 
 st.title("Book of Business Checker")
@@ -58,7 +36,14 @@ with create_tab:
         import_folder='~',
         return_type='analysis',
         sheet_functions=[CHECK],
-        key="recon creation spreadsheet"
+        key="recon creation spreadsheet",
+        code_options={
+            'as_function': True, 
+            'function_name': f'create_{provider_name}_recon',
+            'function_params': [],
+            'call_function': False,
+            'import_custom_python_code': True, 
+        }
     )
 
     if st.button("Save automation"):
